requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
audioread==2.1.6
backcall==0.1.0
bleach==3.0.0
cffi==1.11.5
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
enum34==1.1.6
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
librosa==0.6.3
llvmlite==0.28.0
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numba==0.43.1
numpy==1.15.2
olympic==0.1.5
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
prometheus-client==0.4.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
py==1.7.0
pycparser==2.19
Pygments==2.2.0
pyparsing==2.2.2
PySoundFile==0.9.0.post1
pytest==3.9.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.4.1
resampy==0.2.1
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
